fixed an infinite loop

+++ philo/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: busseven <busseven@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/10 17:54:32 by busseven          #+#    #+#             */
-/*   Updated: 2025/07/12 13:00:40 by busseven         ###   ########.fr       */
+/*   Updated: 2025/07/12 14:58:22 by busseven         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,9 +28,9 @@ typedef struct s_table
 	void				**philo_arr;
 	int					death;
 	int					philo_count;
-	unsigned long long	start_time;
+	size_t				start_time;
 	int					time_to_eat;
-	unsigned long long	time_to_die;
+	size_t				time_to_die;
 	int					time_to_sleep;
 	int					last_param;
 	int					has_last_param;
@@ -41,7 +41,7 @@ typedef struct s_table
 	pthread_mutex_t		*wait_mutex;
 	pthread_mutex_t		*write_mutex;
 	pthread_mutex_t		*full_mutex;
-	unsigned long long	cur_time;
+	size_t				cur_time;
 	int					wait;
 	int					i;
 
@@ -52,12 +52,12 @@ typedef struct s_seat
 	int					num;
 	int					chair_num;
 	int					born_time;
-	unsigned long long	last_eaten;
+	size_t				last_eaten;
 	int					meals_to_eat;
 	int					meals_eaten;
 	int					last_philo;
 	int					has_opt_param;
-	unsigned long long	start_time;
+	size_t				start_time;
 	pthread_t			*philosopher;
 	pthread_mutex_t		*left_fork;
 	pthread_mutex_t		*right_fork;
@@ -78,19 +78,19 @@ void				wrong_format(void);
 void				*ft_calloc(size_t count, size_t size);
 t_seat				*ft_lastnode(t_seat *seats);
 void				prepare_table(t_table *table);
-unsigned long long	get_current_time(void);
+size_t	get_current_time(void);
 int					ft_atoi(const char *str);
 void				set_int(pthread_mutex_t *mtx, int *i, int set);
 int					read_int(pthread_mutex_t *mtx, int *i);
-unsigned long long	read_long(pthread_mutex_t *mtx, unsigned long long *l);
-void				set_longlong(pthread_mutex_t *mtx, unsigned long long *i, unsigned long long set);
-void				ft_putnbr_fd(unsigned long long n, int fd);
+size_t	read_long(pthread_mutex_t *mtx, size_t *l);
+void				set_longlong(pthread_mutex_t *mtx, size_t *i, size_t set);
+void				ft_putnbr_fd(size_t n, int fd);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
 void				write_with_mtx(t_seat *seat, char *action);
-unsigned long long	get_time_stamp(t_seat *seat);
-void				philo_pause(unsigned long long time, int philo_count);
+size_t	get_time_stamp(t_seat *seat);
+void				philo_pause(size_t time, int philo_count);
 void				write_death(t_seat *seat, long long timestamp);
-unsigned long long	time_since_eaten(t_seat *seat);
+size_t	time_since_eaten(t_seat *seat);
 void				*waiter(void *void_table);
 void				*routine(void *void_seat);
